requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.1
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.5.0
opencv-python-headless>=4.5.3
numpy>=1.19.5
python-dotenv==1.0.0
pyngrok>=5.0.0
flask>=2.0.1
transformers>=4.30.0
requests>=2.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
